requirements.txt
requests
pandas
python-dotenv
openpyxl
finance-datareader
pykrx
streamlit
plotly
google-cloud-storage
gcsfs
